MANTA-3525 rename `restart` for `reload`
The old name `restart` says HOW the aim of reloading the HAProxy config was
achieved. With HAProxy seamless config changes the method to reload HAProxy
config is in fact an SMF refresh instead. Rather than changing all reloads to
refreshes I prefer to name the items for what they do and not leak the
abstraction of how (as we may use the HAProxy 2,0 data plane API in future, for
example.)

Also, since SMF refresh is "fire and forget" and does not support the `-s` sync
command, there is no need for the backoff/retry mechanism on the call to refresh.